requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
astunparse==1.6.3
tenacity==8.1.0
openai==1.45.0
numpy==1.26.4
networkx==3.3
matplotlib==3.9.1
pandas==2.2.2
evalplus==0.2.1
tiktoken==0.7.0
anthropic==0.34.2
transformers==4.44.2
scikit-learn==1.5.1
responses==0.18.0
plotly==5.24.0
flask==3.0.3
python-dotenv==1.0.1